requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
ez-setup==0.9
fastjsonschema==2.16.2
Flask==2.2.3
fonttools==4.33.3
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.4
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.3
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.1
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
terminado==0.17.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
urllib3==2.0.7
vboxapi==1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.3
